Oblig1 - ferdig

+++ Oblig 1 - utan oppg2/src/no/hvl/dat108/Tilsett.java
@@ -12,18 +12,30 @@ public class Tilsett {
 	private String fornamn;
 	private String etternamn;
 	private char kjonn;
-	private int arslon;
+	private double arslon;
 	private String stilling;
-	
-	public Tilsett(String fornamn, String etternamn, char kjonn, int arslon, String stilling){
+
+	/**
+	 * Generate constructor
+	 * @param fornamn
+	 * @param etternamn
+	 * @param kjonn
+	 * @param arslon
+	 * @param stilling
+	 */
+	public Tilsett(String fornamn, String etternamn, char kjonn, double arslon, String stilling){
 		this.fornamn = fornamn;
 		this.etternamn = etternamn;
 		this.kjonn = kjonn;
 		this.arslon = arslon;
 		this.stilling = stilling;
 	}
-	
 
+
+	/**
+	 * Generate getters and setters.
+	 *
+	 */
 	public String getFornamn() {
 		return fornamn;
 	}
@@ -45,10 +57,10 @@ public class Tilsett {
 		this.kjonn = kjonn;
 	}
 	
-	public int getArslon() {
+	public double getArslon() {
 		return arslon;
 	}
-	public void setArslon(int arslon) {
+	public void setArslon(double arslon) {
 		this.arslon = arslon;
 	}
 	
@@ -59,12 +71,15 @@ public class Tilsett {
 		this.stilling = stilling;
 	}
 	
-	public int endreLon(Function<Integer, Integer> nyverdi) {
-		return this.arslon = nyverdi.apply(getArslon());
+	public double endreLon(Function<Double, Double> nyverdi) {
+		return this.arslon = nyverdi.apply( getArslon());
 		
 	}
 
 
+	/**
+	 * Generate toString()
+	 */
 	@Override
 	public String toString() {
 		return "Tilsett [fornamn=" + fornamn + ", etternamn=" + etternamn + ", kjonn=" + kjonn + ", arslon=" + arslon
